requirements.txt
pip==22.0.2 
python-telegram-bot==13.13
google-generativeai==0.3.1
numpy
Pillow
huggingface_hub
gradio_client
PyPDF2==3.0.1
python-dotenv
urllib3==1.26.7
onnxruntime
chromadb>=0.4.1
cohere